requirements.txt
anywidget==0.9.21
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
autopep8==2.3.2
# Editable Git install with no remote (backtest_project==0.1)
-e /Users/johnnyf/Documents/Git/backtesting-project
# Editable Git install with no remote (backtesting-project==0.1.0)
-e /Users/johnnyf/Documents/Git/backtesting-project
backtrader==1.9.78.123
black==25.11.0
build==1.3.0
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distlib==0.4.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
identify==2.6.15
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isort==7.0.0
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
learn==1.0.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
# Editable Git install with no remote (myproject==0.1.0)
-e /Users/johnnyf/Documents/Git/backtesting-project
mypy_extensions==1.1.0
narwhals==2.12.0
nbformat==5.10.4
nbqa==1.9.1
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.3.5
oandapy @ git+https://github.com/oanda/oandapy.git@6ad5336b10d4a52bea50ed0ce23b67592c02b2c0
oandapyV20==0.7.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
pre_commit==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
psygnal==0.15.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
Pygments==2.19.2
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
sci-kit==0.1.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
TA-Lib==0.6.8
tokenize_rt==6.2.0
tomli==2.3.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.14
widgetsnbextension==4.0.15
